merge from main br

+++ android_clean_app.py
@@ -13,8 +13,10 @@ import argparse
 import os
 import re
 import subprocess
-import xml.etree.ElementTree as ET
+import distutils.spawn
 from lxml import etree
+
+ANDROID_MANIFEST_FILE = 'AndroidManifest.xml'
 
 
 class Issue:
@@ -22,7 +24,6 @@ class Issue:
     """
     Stores a single issue reported by Android Lint
     """
-    pattern = re.compile('The resource (.+) appears to be unused')
 
     def __init__(self, filepath, remove_file):
         self.filepath = filepath
@@ -30,15 +31,34 @@ class Issue:
         self.elements = []
 
     def __str__(self):
-        return '{0} {1}'.format(self.filepath)
+        return '{0} {1}'.format(self.filepath, self.elements)
 
     def __repr__(self):
-        return '{0} {1}'.format(self.filepath)
+        return '{0} {1}'.format(self.filepath, self.elements)
 
     def add_element(self, message):
-        res = re.findall(Issue.pattern, message)[0]
-        bits = res.split('.')[-2:]
+        res_all = re.findall(self.pattern, message)
+        if res_all:
+            self._process_match(res_all)
+        else:
+            print("The pattern '%s' seems to find nothing in the error message '%s'. We can't find the resource and "
+                  "can't remove it. The pattern might have changed, please check and report this in github issues." % (
+                      self.pattern.pattern, message))
+
+
+class UnusedResourceIssue(Issue):
+    pattern = re.compile('The resource `?([^`]+)`? appears to be unused')
+
+    def _process_match(self, match_result):
+        bits = match_result[0].split('.')[-2:]
         self.elements.append((bits[0], bits[1]))
+
+
+class ExtraTranslationIssue(Issue):
+    pattern = re.compile('The resource string \"`([^`]+)`\" has been marked as `translatable=\"false')
+
+    def _process_match(self, match_result):
+        self.elements.append(('string', match_result[0]))
 
 
 def parse_args():
@@ -54,7 +74,7 @@ def parse_args():
                              'app directory',
                         default='.')
     parser.add_argument('--xml',
-                        help='Path to the list result. If not specifies linting will be done by the script',
+                        help='Path to the lint result. If not specifies linting will be done by the script',
                         default=None)
     parser.add_argument('--ignore-layouts',
                         help='Should ignore layouts',
@@ -69,29 +89,68 @@ def run_lint_command():
     """
     lint, app_dir, lint_result, ignore_layouts = parse_args()
     if not lint_result:
+        if not distutils.spawn.find_executable(lint):
+            raise Exception(
+                '`%s` executable could not be found and path to lint result not specified. See --help' % lint)
         lint_result = os.path.join(app_dir, 'lint-result.xml')
-        call_result = subprocess.call([lint, app_dir, '--xml', lint_result], shell=True)
+        call_result = subprocess.call([lint, app_dir, '--xml', lint_result])
         if call_result > 0:
-            print('Running the command failed. Try running it from the console. Arguments for subprocess.call: {0}'.format(
-                [lint, app_dir, '--xml', lint_result]))
+            print('Running the command failed with result %s. Try running it from the console.'
+                  ' Arguments for subprocess.call: %s' % (call_result, [lint, app_dir, '--xml', lint_result]))
+    else:
+        if not os.path.isabs(lint_result):
+            lint_result = os.path.join(app_dir, lint_result)
+    lint_result = os.path.abspath(lint_result)
     return lint_result, app_dir, ignore_layouts
 
 
-def parse_lint_result(lint_result_path):
-    """
-    Parse lint-result.xml and create Issue for every problem found
-    """
-    root = ET.parse(lint_result_path).getroot()
+def get_manifest_string_refs(manifest_path):
+    pattern = re.compile('="@string/([^"]+)"')
+    with open(manifest_path, 'r') as f:
+        data = f.read()
+        refs = set(re.findall(pattern, data))
+        return [x.replace('/', '.') for x in refs]
+
+
+def _get_issues_from_location(issue_class, locations, message):
     issues = []
+    for location in locations:
+        filepath = location.get('file')
+        # if the location contains line and/or column attribute not the entire resource is unused.
+        # that's a guess ;)
+        # TODO stop guessing
+        remove_entire_file = (location.get('line') or location.get('column')) is None
+        issue = issue_class(filepath, remove_entire_file)
+        issue.add_element(message)
+        issues.append(issue)
+    return issues
+
+
+def parse_lint_result(lint_result_path, manifest_path):
+    """
+    Parse lint-result.xml and create Issue for every problem found except unused strings referenced in AndroidManifest
+    """
+    unused_string_pattern = re.compile('The resource `R\.string\.([^`]+)` appears to be unused')
+    mainfest_string_refs = get_manifest_string_refs(manifest_path)
+    root = etree.parse(lint_result_path).getroot()
+    issues = []
+
     for issue_xml in root.findall('.//issue[@id="UnusedResources"]'):
-        for location in issue_xml.findall('location'):
-            filepath = location.get('file')
-            # if the location contains line and/or column attribute not the entire resource is unused. that's a guess ;)
-            # TODO stop guessing
-            remove_entire_file = (location.get('line') or location.get('column')) is None
-            issue = Issue(filepath, remove_entire_file)
-            issue.add_element(issue_xml.get('message'))
-            issues.append(issue)
+        message = issue_xml.get('message')
+        unused_string = re.match(unused_string_pattern, issue_xml.get('message'))
+        has_string_in_manifest = unused_string and unused_string.group(1) in mainfest_string_refs
+        if not has_string_in_manifest:
+            issues.extend(_get_issues_from_location(UnusedResourceIssue,
+                                                    issue_xml.findall('location'),
+                                                    message))
+
+    for issue_xml in root.findall('.//issue[@id="ExtraTranslation"]'):
+        message = issue_xml.get('message')
+        if re.findall(ExtraTranslationIssue.pattern, message):
+            issues.extend(_get_issues_from_location(ExtraTranslationIssue,
+                                                    issue_xml.findall('location'),
+                                                    message))
+
     return issues
 
 
@@ -108,16 +167,17 @@ def remove_resource_value(issue, filepath):
     """
     Read an xml file and remove an element which is unused, then save the file back to the filesystem
     """
-    for element in issue.elements:
-        print('removing {0} from resource {1}'.format(element, filepath))
-        parser = etree.XMLParser(remove_blank_text=False, remove_comments=False,
-                                 remove_pis=False, strip_cdata=False, resolve_entities=False)
-        tree = etree.parse(filepath, parser)
-        root = tree.getroot()
-        for unused_value in root.findall('.//{0}[@name="{1}"]'.format(element[0], element[1])):
-            root.remove(unused_value)
-        with open(filepath, 'wb') as resource:
-            tree.write(resource, encoding='utf-8', xml_declaration=True)
+    if os.path.exists(filepath):
+        for element in issue.elements:
+            print('removing {0} from resource {1}'.format(element, filepath))
+            parser = etree.XMLParser(remove_blank_text=False, remove_comments=False,
+                                     remove_pis=False, strip_cdata=False, resolve_entities=False)
+            tree = etree.parse(filepath, parser)
+            root = tree.getroot()
+            for unused_value in root.findall('.//{0}[@name="{1}"]'.format(element[0], element[1])):
+                root.remove(unused_value)
+            with open(filepath, 'wb') as resource:
+                tree.write(resource, encoding='utf-8', xml_declaration=True)
 
 
 def remove_unused_resources(issues, app_dir, ignore_layouts):
@@ -134,8 +194,12 @@ def remove_unused_resources(issues, app_dir, ignore_layouts):
 
 def main():
     lint_result_path, app_dir, ignore_layouts = run_lint_command()
-    issues = parse_lint_result(lint_result_path)
-    remove_unused_resources(issues, app_dir, ignore_layouts)
+    if os.path.exists(lint_result_path):
+        manifest_path = os.path.abspath(os.path.join(app_dir, ANDROID_MANIFEST_FILE))
+        issues = parse_lint_result(lint_result_path, manifest_path)
+        remove_unused_resources(issues, app_dir, ignore_layouts)
+    else:
+        print('the file with lint results could not be found: %s' % lint_result_path)
 
 
 if __name__ == '__main__':
